Refactor scope creation dialog and plans management
- Moved plan selection UI into a new InstallCard component for better separation of concerns.
- Updated plans data to use localized strings for tags.
- Simplified the sheet-create-scope component by integrating a new CreateScope form.
- Enhanced the scopes page to include a button for creating new scopes, with a dialog for the creation process.
- Removed unnecessary state management and API calls from the sheet-create-scope component.
