+++ web/src/routes/(auth)/scopes/+page.svelte
@@ -6,13 +6,14 @@
 	import { resolve } from '$app/paths';
 	import { type Scope, ScopeService } from '$lib/api/scope/v1/scope_pb';
 	import SquareGridImage from '$lib/assets/square-grid.svg';
+	import DialogCreateScope from '$lib/components/layout/dialog-create-scope.svelte';
 	import { scopeIcon } from '$lib/components/scopes/icon';
 	import { Badge } from '$lib/components/ui/badge';
 	import { Button } from '$lib/components/ui/button';
 	import * as Card from '$lib/components/ui/card';
 	import { m } from '$lib/paraglide/messages';
 
-	const EXCLUDED_SCOPES = ['cos', 'cos-dev', 'cos-lite'];
+	const EXCLUDED_SCOPES = ['cos', 'cos-dev', 'cos-lite', 'aaa', 'bbb', 'ccc', 'zzz'];
 
 	const transport: Transport = getContext('transport');
 	const scopeClient = createClient(ScopeService, transport);
@@ -40,6 +41,7 @@
 		return scopes.findIndex((scope) => scope.name === scopeName);
 	}
 
+	let open = $state(false);
 	let mounted = $state(false);
 	onMount(async () => {
 		await fetchScopes();
@@ -65,7 +67,9 @@
 
 	<!-- Header -->
 	{#if mounted}
-		<h2 class="text-center text-3xl font-bold tracking-tight sm:text-4xl">{m.scope_selector()}</h2>
+		<h2 class="text-center text-3xl font-bold tracking-tight sm:text-4xl">
+			{m.scope_selector()}
+		</h2>
 		<p class="mt-4 text-center text-lg text-muted-foreground">{m.scope_selector_description()}</p>
 	{:else}
 		<h2 class="text-center text-3xl font-bold tracking-tight sm:text-4xl">
@@ -81,42 +85,28 @@
 			{m.scope_selector_loading_description()}
 		</Button>
 	{/if}
-
 	<!-- Scopes Grid -->
 	<div class="z-10 mx-auto grid w-full grid-cols-8 gap-4 px-4 py-10 sm:gap-6 xl:px-0 2xl:w-3/4">
-		{#each scopes as scope, index (scope.name)}
-			<a
-				href={resolve(`/(auth)/scope/[scope]`, { scope: scope.name })}
-				class="group col-span-2 cursor-pointer {getCardColumnClass(index, scopes.length)}"
-			>
-				<Card.Root>
+		{#if scopes.length === 0}
+			<!-- Add Scope Card -->
+			<button onclick={() => (open = true)} class="group col-span-2 col-start-4 cursor-pointer">
+				<Card.Root class="transition-all duration-200 hover:scale-105 hover:shadow-lg">
 					<Card.Header class="gap-0">
 						<div class="flex items-center gap-4">
-							<!-- Scope Icon -->
+							<!-- Add Icon -->
 							<div class="flex size-10 items-center justify-center rounded-lg bg-primary">
-								<Icon
-									icon="{scopeIcon(getScopeIndex(scope.name))}-fill"
-									class="size-6 text-primary-foreground"
-								/>
+								<Icon icon="ph:plus-bold" class="size-6 text-card" />
 							</div>
 
-							<!-- Scope Info -->
+							<!-- Add Scope Info -->
 							<div class="grid -space-y-1">
-								<Card.Description class="capitalize">
-									{scope.status}
-								</Card.Description>
-								<Card.Title class="text-2xl text-nowrap">{scope.name}</Card.Title>
+								<Card.Description class="capitalize">{m.create()}</Card.Description>
+								<Card.Title class="text-2xl text-nowrap">New Scope</Card.Title>
 							</div>
 						</div>
 
-						<!-- Scope Stats and Action -->
+						<!-- Action -->
 						<Card.Action class="overflow-hidden group-hover:self-center">
-							<Badge variant="outline" class="hidden group-hover:hidden lg:block">
-								<span class="text-green-600">
-									{m.machines()}: {scope.machineCount > 0 ? scope.machineCount : '-'} /
-									{m.unit()}: {scope.unitCount > 0 ? scope.unitCount : '-'}
-								</span>
-							</Badge>
 							<div
 								class="hidden size-10 items-center justify-center rounded-full text-muted-foreground transition-colors group-hover:inline-flex group-hover:bg-primary group-hover:text-primary-foreground"
 							>
@@ -125,7 +115,52 @@
 						</Card.Action>
 					</Card.Header>
 				</Card.Root>
-			</a>
-		{/each}
+			</button>
+			<DialogCreateScope bind:open />
+		{:else}
+			{#each scopes as scope, index (scope.name)}
+				<a
+					href={resolve(`/(auth)/scope/[scope]`, { scope: scope.name })}
+					class="group col-span-2 cursor-pointer {getCardColumnClass(index, scopes.length)}"
+				>
+					<Card.Root>
+						<Card.Header class="gap-0">
+							<div class="flex items-center gap-4">
+								<!-- Scope Icon -->
+								<div class="flex size-10 items-center justify-center rounded-lg bg-primary">
+									<Icon
+										icon="{scopeIcon(getScopeIndex(scope.name))}-fill"
+										class="size-6 text-primary-foreground"
+									/>
+								</div>
+
+								<!-- Scope Info -->
+								<div class="grid -space-y-1">
+									<Card.Description class="capitalize">
+										{scope.status}
+									</Card.Description>
+									<Card.Title class="text-2xl text-nowrap">{scope.name}</Card.Title>
+								</div>
+							</div>
+
+							<!-- Scope Stats and Action -->
+							<Card.Action class="overflow-hidden group-hover:self-center">
+								<Badge variant="outline" class="hidden group-hover:hidden lg:block">
+									<span class="text-green-600">
+										{m.machines()}: {scope.machineCount > 0 ? scope.machineCount : '-'} /
+										{m.unit()}: {scope.unitCount > 0 ? scope.unitCount : '-'}
+									</span>
+								</Badge>
+								<div
+									class="hidden size-10 items-center justify-center rounded-full text-muted-foreground transition-colors group-hover:inline-flex group-hover:bg-primary group-hover:text-primary-foreground"
+								>
+									<Icon icon="ph:arrow-right" class="size-6" />
+								</div>
+							</Card.Action>
+						</Card.Header>
+					</Card.Root>
+				</a>
+			{/each}
+		{/if}
 	</div>
 </main>
